// src/features/bubble/types.ts
export type BubbleParams = {
  theme?: BubbleTheme;
};

export type BubbleTheme = {
  poweredByVisibility?: boolean;
  header?: {
    title?: string;
    subTitle?: string;
    backgroundColor?: string;
    textColor?: string;
    avatar?: string;
  };
  loginPrompt: {
    field_name: string;
    is_required: boolean;
  }[];
  login_prompt?: {
    field_name: string;
    is_required: boolean;
  }[];
  chatWindow?: ChatWindowTheme;
  button?: ButtonTheme;
};

export type TextInputTheme = {
  backgroundColor?: string;
  textColor?: string;
  placeholder?: string;
  sendButtonColor?: string;
};

export type UserMessageTheme = {
  backgroundColor?: string;
  textColor?: string;
  showAvatar?: boolean;
  avatarSrc?: string;
};

export type BotMessageTheme = {
  backgroundColor?: string;
  textColor?: string;
  showAvatar?: boolean;
  avatarSrc?: string;
};

export type ChatWindowTheme = {
  welcomeMessage?: string;
  backgroundColor?: string;
  height?: number;
  width?: number;
  fontSize?: number;
  userMessage?: UserMessageTheme;
  botMessage?: BotMessageTheme;
  textInput?: TextInputTheme;
  poweredByTextColor?: string;
};

export type ButtonTheme = {
  size?: "medium" | "large";
  backgroundColor?: string;
  iconColor?: string;
  customIconSrc?: string;
  bottom?: number;
  right?: number;
};